Conflicts:
	src/Basescape/CraftSoldiersState.cpp
	src/Battlescape/BattlescapeState.cpp
	src/Menu/LoadGameState.cpp
	src/Menu/OptionsVideoState.cpp

Inactive FPS disabled for now, should be enabled in next commits.

+++ src/Basescape/CraftSoldiersState.cpp
@@ -176,7 +176,7 @@ void CraftSoldiersState::init()
  */
 void CraftSoldiersState::lstItemsLeftArrowClick(Action *action)
 {
-	int row = _lstSoldiers->getSelectedRow();
+	unsigned int row = _lstSoldiers->getSelectedRow();
 	if (row > 0)
 	{
 		if (action->getDetails()->button.button == SDL_BUTTON_LEFT)
@@ -196,7 +196,7 @@ void CraftSoldiersState::lstItemsLeftArrowClick(Action *action)
  * @param row Selected soldier row.
  * @param max Move the soldier to the top?
  */
-void CraftSoldiersState::moveSoldierUp(Action *action, int row, bool max)
+void CraftSoldiersState::moveSoldierUp(Action *action, unsigned int row, bool max)
 {
 	Soldier *s = _base->getSoldiers()->at(row);
 	if (max)
@@ -227,9 +227,9 @@ void CraftSoldiersState::moveSoldierUp(Action *action, int row, bool max)
  */
 void CraftSoldiersState::lstItemsRightArrowClick(Action *action)
 {
-	int row = _lstSoldiers->getSelectedRow();
+	unsigned int row = _lstSoldiers->getSelectedRow();
 	size_t numSoldiers = _base->getSoldiers()->size();
-	if (0 < numSoldiers && INT_MAX >= numSoldiers && row < (int)numSoldiers - 1)
+	if (0 < numSoldiers && INT_MAX >= numSoldiers && row < numSoldiers - 1)
 	{
 		if (action->getDetails()->button.button == SDL_BUTTON_LEFT)
 		{
@@ -248,7 +248,7 @@ void CraftSoldiersState::lstItemsRightArrowClick(Action *action)
  * @param row Selected soldier row.
  * @param max Move the soldier to the bottom?
  */
-void CraftSoldiersState::moveSoldierDown(Action *action, int row, bool max)
+void CraftSoldiersState::moveSoldierDown(Action *action, unsigned int row, bool max)
 {
 	Soldier *s = _base->getSoldiers()->at(row);
 	if (max)
@@ -336,7 +336,7 @@ void CraftSoldiersState::lstSoldiersMouseWheel(Action *action)
 {
 	if (Options::changeValueByMouseWheel == 0)
 		return;
-	int row = _lstSoldiers->getSelectedRow();
+	unsigned int row = _lstSoldiers->getSelectedRow();
 	size_t numSoldiers = _base->getSoldiers()->size();
 	const SDL_Event &ev(*action->getDetails());
 	if (ev.type == SDL_MOUSEWHEEL)
@@ -360,6 +360,5 @@ void CraftSoldiersState::lstSoldiersMouseWheel(Action *action)
 		}
 	}
 }
-
 
 }
